backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
fpdf2
httpx
